Struct hassle_rs::intellisense::DxcCursorKind

pub struct DxcCursorKind { /* private fields */ }

Implementations

impl DxcCursorKind

pub const UNEXPOSED_DECL: Self = _

pub const STRUCT_DECL: Self = _

pub const UNION_DECL: Self = _

pub const CLASS_DECL: Self = _

pub const ENUM_DECL: Self = _

pub const FIELD_DECL: Self = _

pub const ENUM_CONSTANT_DECL: Self = _

pub const FUNCTION_DECL: Self = _

pub const VAR_DECL: Self = _

pub const PARM_DECL: Self = _

pub const OBJ_C_INTERFACE_DECL: Self = _

pub const OBJ_C_CATEGORY_DECL: Self = _

pub const OBJ_C_PROTOCOL_DECL: Self = _

pub const OBJ_C_PROPERTY_DECL: Self = _

pub const OBJ_C_IVAR_DECL: Self = _

pub const OBJ_C_INSTANCE_METHOD_DECL: Self = _

pub const OBJ_C_CLASS_METHOD_DECL: Self = _

pub const OBJ_C_IMPLEMENTATION_DECL: Self = _

pub const OBJ_C_CATEGORY_IMPL_DECL: Self = _

pub const TYPEDEF_DECL: Self = _

pub const CXX_METHOD: Self = _

pub const NAMESPACE: Self = _

pub const LINKAGE_SPEC: Self = _

pub const CONSTRUCTOR: Self = _

pub const DESTRUCTOR: Self = _

pub const CONVERSION_FUNCTION: Self = _

pub const TEMPLATE_TYPE_PARAMETER: Self = _

pub const NON_TYPE_TEMPLATE_PARAMETER: Self = _

pub const TEMPLATE_TEMPLATE_PARAMETER: Self = _

pub const FUNCTION_TEMPLATE: Self = _

pub const CLASS_TEMPLATE: Self = _

pub const CLASS_TEMPLATE_PARTIAL_SPECIALIZATION: Self = _

pub const NAMESPACE_ALIAS: Self = _

pub const USING_DIRECTIVE: Self = _

pub const USING_DECLARATION: Self = _

pub const TYPE_ALIAS_DECL: Self = _

pub const OBJ_C_SYNTHESIZE_DECL: Self = _

pub const OBJ_C_DYNAMIC_DECL: Self = _

pub const CXX_ACCESS_SPECIFIER: Self = _

pub const FIRST_DECL: Self = _

pub const LAST_DECL: Self = _

pub const FIRST_REF: Self = _

pub const OBJ_C_SUPER_CLASS_REF: Self = _

pub const OBJ_C_PROTOCOL_REF: Self = _

pub const OBJ_C_CLASS_REF: Self = _

pub const TYPE_REF: Self = _

pub const CXX_BASE_SPECIFIER: Self = _

pub const TEMPLATE_REF: Self = _

pub const NAMESPACE_REF: Self = _

pub const MEMBER_REF: Self = _

pub const LABEL_REF: Self = _

pub const OVERLOADED_DECL_REF: Self = _

pub const VARIABLE_REF: Self = _

pub const LAST_REF: Self = _

pub const FIRST_INVALID: Self = _

pub const INVALID_FILE: Self = _

pub const NO_DECL_FOUND: Self = _

pub const NOT_IMPLEMENTED: Self = _

pub const INVALID_CODE: Self = _

pub const LAST_INVALID: Self = _

pub const FIRST_EXPR: Self = _

pub const UNEXPOSED_EXPR: Self = _

pub const DECL_REF_EXPR: Self = _

pub const MEMBER_REF_EXPR: Self = _

pub const CALL_EXPR: Self = _

pub const OBJ_C_MESSAGE_EXPR: Self = _

pub const BLOCK_EXPR: Self = _

pub const INTEGER_LITERAL: Self = _

pub const FLOATING_LITERAL: Self = _

pub const IMAGINARY_LITERAL: Self = _

pub const STRING_LITERAL: Self = _

pub const CHARACTER_LITERAL: Self = _

pub const PAREN_EXPR: Self = _

pub const UNARY_OPERATOR: Self = _

pub const ARRAY_SUBSCRIPT_EXPR: Self = _

pub const BINARY_OPERATOR: Self = _

pub const COMPOUND_ASSIGN_OPERATOR: Self = _

pub const CONDITIONAL_OPERATOR: Self = _

pub const C_STYLE_CAST_EXPR: Self = _

pub const COMPOUND_LITERAL_EXPR: Self = _

pub const INIT_LIST_EXPR: Self = _

pub const ADDR_LABEL_EXPR: Self = _

pub const STMT_EXPR: Self = _

pub const GENERIC_SELECTION_EXPR: Self = _

pub const GNU_NULL_EXPR: Self = _

pub const CXX_STATIC_CAST_EXPR: Self = _

pub const CXX_DYNAMIC_CAST_EXPR: Self = _

pub const CXX_REINTERPRET_CAST_EXPR: Self = _

pub const CXX_CONST_CAST_EXPR: Self = _

pub const CXX_FUNCTIONAL_CAST_EXPR: Self = _

pub const CXX_TYPEID_EXPR: Self = _

pub const CXX_BOOL_LITERAL_EXPR: Self = _

pub const CXX_NULL_PTR_LITERAL_EXPR: Self = _

pub const CXX_THIS_EXPR: Self = _

pub const CXX_THROW_EXPR: Self = _

pub const CXX_NEW_EXPR: Self = _

pub const CXX_DELETE_EXPR: Self = _

pub const UNARY_EXPR: Self = _

pub const OBJ_C_STRING_LITERAL: Self = _

pub const OBJ_C_ENCODE_EXPR: Self = _

pub const OBJ_C_SELECTOR_EXPR: Self = _

pub const OBJ_C_PROTOCOL_EXPR: Self = _

pub const OBJ_C_BRIDGED_CAST_EXPR: Self = _

pub const PACK_EXPANSION_EXPR: Self = _

pub const SIZE_OF_PACK_EXPR: Self = _

pub const LAMBDA_EXPR: Self = _

pub const OBJ_C_BOOL_LITERAL_EXPR: Self = _

pub const OBJ_C_SELF_EXPR: Self = _

pub const LAST_EXPR: Self = _

pub const FIRST_STMT: Self = _

pub const UNEXPOSED_STMT: Self = _

pub const LABEL_STMT: Self = _

pub const COMPOUND_STMT: Self = _

pub const CASE_STMT: Self = _

pub const DEFAULT_STMT: Self = _

pub const IF_STMT: Self = _

pub const SWITCH_STMT: Self = _

pub const WHILE_STMT: Self = _

pub const DO_STMT: Self = _

pub const FOR_STMT: Self = _

pub const GOTO_STMT: Self = _

pub const INDIRECT_GOTO_STMT: Self = _

pub const CONTINUE_STMT: Self = _

pub const BREAK_STMT: Self = _

pub const RETURN_STMT: Self = _

pub const GCC_ASM_STMT: Self = _

pub const ASM_STMT: Self = _

pub const OBJ_C_AT_TRY_STMT: Self = _

pub const OBJ_C_AT_CATCH_STMT: Self = _

pub const OBJ_C_AT_FINALLY_STMT: Self = _

pub const OBJ_C_AT_THROW_STMT: Self = _

pub const OBJ_C_AT_SYNCHRONIZED_STMT: Self = _

pub const OBJ_C_AUTORELEASE_POOL_STMT: Self = _

pub const OBJ_C_FOR_COLLECTION_STMT: Self = _

pub const CXX_CATCH_STMT: Self = _

pub const CXX_TRY_STMT: Self = _

pub const CXX_FOR_RANGE_STMT: Self = _

pub const SEH_TRY_STMT: Self = _

pub const SEH_EXCEPT_STMT: Self = _

pub const SEH_FINALLY_STMT: Self = _

pub const MS_ASM_STMT: Self = _

pub const NULL_STMT: Self = _

pub const DECL_STMT: Self = _

pub const OMP_PARALLEL_DIRECTIVE: Self = _

pub const OMP_SIMD_DIRECTIVE: Self = _

pub const OMP_FOR_DIRECTIVE: Self = _

pub const OMP_SECTIONS_DIRECTIVE: Self = _

pub const OMP_SECTION_DIRECTIVE: Self = _

pub const OMP_SINGLE_DIRECTIVE: Self = _

pub const OMP_PARALLEL_FOR_DIRECTIVE: Self = _

pub const OMP_PARALLEL_SECTIONS_DIRECTIVE: Self = _

pub const OMP_TASK_DIRECTIVE: Self = _

pub const OMP_MASTER_DIRECTIVE: Self = _

pub const OMP_CRITICAL_DIRECTIVE: Self = _

pub const OMP_TASKYIELD_DIRECTIVE: Self = _

pub const OMP_BARRIER_DIRECTIVE: Self = _

pub const OMP_TASKWAIT_DIRECTIVE: Self = _

pub const OMP_FLUSH_DIRECTIVE: Self = _

pub const SEH_LEAVE_STMT: Self = _

pub const OMP_ORDERED_DIRECTIVE: Self = _

pub const OMP_ATOMIC_DIRECTIVE: Self = _

pub const OMP_FOR_SIMD_DIRECTIVE: Self = _

pub const OMP_PARALLEL_FOR_SIMD_DIRECTIVE: Self = _

pub const OMP_TARGET_DIRECTIVE: Self = _

pub const OMP_TEAMS_DIRECTIVE: Self = _

pub const OMP_TASKGROUP_DIRECTIVE: Self = _

pub const OMP_CANCELLATION_POINT_DIRECTIVE: Self = _

pub const OMP_CANCEL_DIRECTIVE: Self = _

pub const LAST_STMT: Self = _

pub const TRANSLATION_UNIT: Self = _

pub const FIRST_ATTR: Self = _

pub const UNEXPOSED_ATTR: Self = _

pub const IB_ACTION_ATTR: Self = _

pub const IB_OUTLET_ATTR: Self = _

pub const IB_OUTLET_COLLECTION_ATTR: Self = _

pub const CXX_FINAL_ATTR: Self = _

pub const CXX_OVERRIDE_ATTR: Self = _

pub const ANNOTATE_ATTR: Self = _

pub const ASM_LABEL_ATTR: Self = _

pub const PACKED_ATTR: Self = _

pub const PURE_ATTR: Self = _

pub const CONST_ATTR: Self = _

pub const NO_DUPLICATE_ATTR: Self = _

pub const CUDA_CONSTANT_ATTR: Self = _

pub const CUDA_DEVICE_ATTR: Self = _

pub const CUDA_GLOBAL_ATTR: Self = _

pub const CUDA_HOST_ATTR: Self = _

pub const CUDA_SHARED_ATTR: Self = _

pub const LAST_ATTR: Self = _

pub const PREPROCESSING_DIRECTIVE: Self = _

pub const MACRO_DEFINITION: Self = _

pub const MACRO_EXPANSION: Self = _

pub const MACRO_INSTANTIATION: Self = _

pub const INCLUSION_DIRECTIVE: Self = _

pub const FIRST_PREPROCESSING: Self = _

pub const LAST_PREPROCESSING: Self = _

pub const MODULE_IMPORT_DECL: Self = _

pub const FIRST_EXTRA_DECL: Self = _

pub const LAST_EXTRA_DECL: Self = _

pub const fn empty() -> Self

Returns an empty set of flags.

pub const fn all() -> Self

Returns the set containing all flags.

pub const fn bits(&self) -> u32

Returns the raw value of the flags currently stored.

pub const fn from_bits(bits: u32) -> Option<Self>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub const fn from_bits_truncate(bits: u32) -> Self

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub const unsafe fn from_bits_unchecked(bits: u32) -> Self

Convert from underlying bit representation, preserving all bits (even those not corresponding to a defined flag).

Safety

The caller of the bitflags! macro can chose to allow or disallow extra bits for their bitflags type.

The caller of from_bits_unchecked() has to ensure that all bits correspond to a defined flag or that extra bits are valid for this bitflags type.

pub const fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub const fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub const fn intersects(&self, other: Self) -> bool

Returns true if there are flags common to both self and other.

pub const fn contains(&self, other: Self) -> bool

Returns true if all of the flags in other are contained within self.

pub fn insert(&mut self, other: Self)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: Self)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: Self)

Toggles the specified flags in-place.

pub fn set(&mut self, other: Self, value: bool)

Inserts or removes the specified flags depending on the passed value.

pub const fn intersection(self, other: Self) -> Self

Returns the intersection between the flags in self and other.

Specifically, the returned set contains only the flags which are present in both self and other.

This is equivalent to using the & operator (e.g. ops::BitAnd), as in flags & other.

pub const fn union(self, other: Self) -> Self

Returns the union of between the flags in self and other.

Specifically, the returned set contains all flags which are present in either self or other, including any which are present in both (see Self::symmetric_difference if that is undesirable).

This is equivalent to using the | operator (e.g. ops::BitOr), as in flags | other.

pub const fn difference(self, other: Self) -> Self

Returns the difference between the flags in self and other.

Specifically, the returned set contains all flags present in self, except for the ones present in other.

It is also conceptually equivalent to the “bit-clear” operation: flags & !other (and this syntax is also supported).

This is equivalent to using the - operator (e.g. ops::Sub), as in flags - other.

pub const fn symmetric_difference(self, other: Self) -> Self

Returns the symmetric difference between the flags in self and other.

Specifically, the returned set contains the flags present which are present in self or other, but that are not present in both. Equivalently, it contains the flags present in exactly one of the sets self and other.

This is equivalent to using the ^ operator (e.g. ops::BitXor), as in flags ^ other.

pub const fn complement(self) -> Self

Returns the complement of this set of flags.

Specifically, the returned set contains all the flags which are not set in self, but which are allowed for this type.

Alternatively, it can be thought of as the set difference between Self::all() and self (e.g. Self::all() - self)

This is equivalent to using the ! operator (e.g. ops::Not), as in !flags.